+++ test_files/config_8Ch_tcp_relay/launch_gui.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/bash
+
+# ============================================================
+# Launch ADOdin EDM GUI
+# ============================================================
+
+/odin/epics/support/ADOdin/iocs/xspress/bin/linux-x86_64/stxspress-gui