+++ test_files/config_8Ch_tcp_relay/stTcpRelay.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# TCP relay server to fan out the TCP traffic from the Xspress system
+/odin/tools/TcpRelay -s $1 -c $2